requirements.txt
asgiref==3.3.4
auth==0.5.3
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
djangorestframework==3.12.4
dnspython==1.16.0
drf-yasg==1.20.0
eventlet==0.31.0
falcon==3.0.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
packaging==20.9
psycopg2-binary==2.8.6
pymongo==3.11.4
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.5
ruamel.yaml.clib==0.2.2
serializers==0.2.4
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
django-extensions
